finished debug about tage

+++ src/bct.c
@@ -23,7 +23,7 @@ void BCT_Initial(BCT* BranchChooserTable, uint32_t index_width)
 		BranchChooserTable->chooser[i] = weakly_bimodal;
 }
 
-Predictor BCT_Predict(BCT* BranchChooserTable, uint32_t addr)
+Predictor BCT_Predict(BCT* BranchChooserTable, uint64_t addr)
 {
 	uint32_t index = Get_Index(addr, BranchChooserTable->attributes.index_width);
 	switch (BranchChooserTable->chooser[index])
@@ -36,7 +36,7 @@ Predictor BCT_Predict(BCT* BranchChooserTable, uint32_t addr)
 	}
 }
 
-void BCT_Update(BCT* BranchChooserTable, uint32_t addr, Result result)
+void BCT_Update(BCT* BranchChooserTable, uint64_t addr, Result result)
 {
 	if (result.actual_taken == result.predict_taken[BIMODAL] && result.actual_taken == result.predict_taken[GSHARE])
 		return;
